+++ src/poll/controllers/AddAnswersController.java
@@ -3,6 +3,8 @@ package poll.controllers;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
+import java.rmi.RemoteException;
+
 import javax.swing.JTextField;
 
 import poll.model.Poll;
@@ -18,6 +20,10 @@ public class AddAnswersController implements ActionListener {
 	@Override
 	public void actionPerformed(ActionEvent e) {
 		JTextField textField = (JTextField) e.getSource();
-		model.addAnswer(textField.getText());
+		try {
+			model.addAnswer(textField.getText());
+		} catch (RemoteException re) {
+			re.printStackTrace();
+		}
 	}
 }
